# cloud/openapi/shakecloud.yaml
openapi: 3.1.0
...
info:
  title: shake-cloud API
  version: 0.1.0
  description: |
    Login, access keys, the audit log and compute. Instances can be created
    with free-form cpu, memory and disk; named instance types are only a
    shorthand starting point, not a fixed catalogue of sizes. Quotas and
    capacity limits are administrator-settable at runtime through
    /v1/limits, not baked into the deployment.

    Browsers log in through Authentik and hold a session cookie. Terraform
    and the CLI send `Authorization: Bearer sca_<access_key_id>.<secret>`.
    Keys work while Authentik is down.
servers:
  - url: http://192.168.10.205:8080
security:
  - accessKey: []
  - session: []
...
paths:
...
  /auth/login:
    get:
      operationId: StartLogin
      tags: [auth]
      summary: Redirect the browser to Authentik.
      security: []
      responses:
        '302':
          description: Redirect to the Authentik authorization endpoint. Sets a short-lived login cookie bound to the state.
        '503':
          description: Authentik could not be reached. Access keys are unaffected.
  /auth/callback:
    get:
      operationId: CompleteLogin
      tags: [auth]
      summary: Finish the OIDC login and start a session.
      description: |
        Requires membership in cloud-users or cloud-admins. An unknown subject
        carrying the email of an existing account is refused rather than given
        a second account.
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, required: true, schema: {type: string}}
      responses:
        '302':
          description: Logged in; redirect to the portal. Sets the session cookie.
        '400':
          description: State missing, reused, expired or not bound to this browser.
        '403':
          description: Not a member of the cloud groups, or the email belongs to another account.
        '502':
          description: Authentik returned an unusable token.
...
  /v1/caller-identity:
    get:
      operationId: GetCallerIdentity
      tags: [iam]
      summary: Return who the credential belongs to.
      description: Like STS GetCallerIdentity. Useful to check a key before running Terraform.
      responses:
        '200':
          description: The caller.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CallerIdentity'}
        '401': {$ref: '#/components/responses/Error'}
  /v1/access-keys:
    get:
      operationId: ListAccessKeys
      tags: [iam]
      summary: List the caller's access keys that have not been deleted.
      responses:
        '200':
          description: Keys, oldest first. Secrets are never returned.
          content:
            application/json:
              schema:
                type: object
                required: [access_keys]
                properties:
                  access_keys:
                    type: array
                    items: {$ref: '#/components/schemas/AccessKey'}
        '401': {$ref: '#/components/responses/Error'}
    post:
      operationId: CreateAccessKey
      tags: [iam]
      summary: Create an access key. Portal session only.
      description: |
        An access key cannot create another access key, so a leaked key cannot
        make itself permanent. The secret is returned in this response only.
        An account may hold 5 active keys.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                description: {type: string, maxLength: 256}
                expires_in_days:
                  type: integer
                  minimum: 1
                  maximum: 3650
                  description: Omit for a key that does not expire.
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                type: object
                required: [access_key, secret_access_key]
                properties:
                  access_key: {$ref: '#/components/schemas/AccessKey'}
                  secret_access_key:
                    type: string
                    pattern: '^sca_[a-z2-7]{20}\.[A-Za-z0-9_-]{43}$'
                    description: The complete bearer credential. Shown once.
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /v1/access-keys/{access_key_id}:
    delete:
      operationId: DeleteAccessKey
      tags: [iam]
      summary: Delete (revoke) an access key.
      description: |
        Owners may delete their own keys; cloud-admins may delete anyone's.
        The row is kept, inactive, so the audit log still names it.
      parameters:
        - name: access_key_id
          in: path
          required: true
          schema: {type: string, pattern: '^[a-z2-7]{20}$'}
      responses:
        '204':
          description: Deleted.
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /v1/key-pairs:
    get:
      operationId: DescribeKeyPairs
      tags: [iam]
      summary: List the caller's SSH keys.
      description: A caller sees only their own.
      responses:
        '200':
          description: Key pairs.
          content:
            application/json:
              schema:
                type: object
                required: [key_pairs]
                properties:
                  key_pairs:
                    type: array
                    items: {$ref: '#/components/schemas/KeyPair'}
        '401': {$ref: '#/components/responses/Error'}
    post:
      operationId: ImportKeyPair
      tags: [iam]
      summary: Register an SSH public key.
      description: |
        Only the **public** key is stored; the API never holds a private
        key. The key is written into an instance's seed image at launch
        when RunInstances names it, so deleting a key pair later does not
        lock anyone out of an existing VM. Key names are unique per
        account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [key_name, public_key]
              properties:
                key_name: {type: string, maxLength: 64}
                public_key: {type: string, maxLength: 16384}
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KeyPairEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/key-pairs/{key_name}:
    delete:
      operationId: DeleteKeyPair
      tags: [iam]
      summary: Delete one of the caller's SSH keys.
      parameters:
        - name: key_name
          in: path
          required: true
          schema: {type: string, maxLength: 64}
      responses:
        '204':
          description: Deleted.
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /v1/audit-events:
    get:
      operationId: LookupEvents
      tags: [audit]
      summary: Read the audit log, newest first.
      description: |
        Records every change, every login, and failed attempts with a real
        access key ID. Users see their own account; cloud-admins see all.
        event_name is the operationId of the request.
      parameters:
        - name: account_id
          in: query
          description: cloud-admins only.
          schema: {type: string, pattern: '^[0-9]{12}$'}
        - {name: access_key_id, in: query, schema: {type: string}}
        - {name: event_name, in: query, schema: {type: string}}
        - name: max_results
          in: query
          schema: {type: integer, minimum: 1, maximum: 1000, default: 50}
        - name: next_token
          in: query
          description: next_token from the previous page.
          schema: {type: string}
      responses:
        '200':
          description: Events.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: {$ref: '#/components/schemas/AuditEvent'}
                  next_token:
                    type: string
                    description: Present when more events may follow.
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /v1/images:
    get:
      operationId: DescribeImages
      tags: [compute]
      summary: List the deployment's shared images and every account's uploaded ones.
      description: 'So people can see what is available to launch from, and who added it.'
      responses:
        '200':
          description: Images.
          content:
            application/json:
              schema:
                type: object
                required: [images]
                properties:
                  images:
                    type: array
                    items: {$ref: '#/components/schemas/Image'}
        '401': {$ref: '#/components/responses/Error'}
    post:
      operationId: ImportImage
      tags: [compute]
      summary: Upload a disk image.
      description: |
        The image is streamed straight through to the node's image store,
        never buffered in the API, because the API container has only a
        small RAM-backed /tmp. Proxmox validates the file with qemu-img, so
        a corrupt or unsupported file is refused and nothing is stored.
        qcow2 and raw are what the store accepts.

        The request is answered when the store has accepted the image, so a
        large upload holds the connection open for minutes.

        The size is bounded by the administrator-settable
        `capacity.max_image_gib` limit and by the free space the image store
        must keep (`capacity.image_store_min_free_mib`).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name: {type: string, maxLength: 128}
                file: {type: string, format: binary}
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: Uploaded.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ImageEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/images/{image_id}:
    delete:
      operationId: DeleteImage
      tags: [compute]
      summary: Delete an uploaded image.
      description: |
        An owner or a cloud-admin may delete an uploaded image; the
        deployment's **shared** images cannot be deleted through the API
        (they are declared in Terraform). Instances already created from an
        image are unaffected, because their disk was copied at creation. A
        deletion is refused while an instance that was launched from it is
        still being created.
      parameters:
        - $ref: '#/components/parameters/ImageID'
      responses:
        '204':
          description: Deleted.
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/isos:
    get:
      operationId: DescribeISOs
      tags: [compute]
      summary: List every account's uploaded installation ISOs.
      description: 'ISO media is attached as a CD-ROM; it never becomes a root disk. Everyone sees the list, as with images.'
      responses:
        '200':
          description: ISOs.
          content:
            application/json:
              schema:
                type: object
                required: [isos]
                properties:
                  isos:
                    type: array
                    items: {$ref: '#/components/schemas/ISO'}
        '401': {$ref: '#/components/responses/Error'}
    post:
      operationId: ImportISO
      tags: [compute]
      summary: Upload an installation ISO.
      description: |
        The ISO is streamed to the node's image store under `:iso/`, never
        buffered in the API. It is attached as a CD-ROM when an instance is
        created with `install_iso_id`.

        The size is bounded by `capacity.max_image_gib` and by the free space
        the image store must keep (`capacity.image_store_min_free_mib`).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name: {type: string, maxLength: 128}
                os:
                  type: string
                  enum: ['', linux, windows]
                  description: 'windows selects the Windows 11 virtual hardware when this ISO is installed.'
                file: {type: string, format: binary}
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: Uploaded.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ISOEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/isos/{iso_id}:
    delete:
      operationId: DeleteISO
      tags: [compute]
      summary: Delete an uploaded ISO.
      description: An owner or a cloud-admin may delete it. Refused while a live instance still attaches it.
      parameters:
        - $ref: '#/components/parameters/ISOID'
      responses:
        '204':
          description: Deleted.
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/instance-types:
    get:
      operationId: DescribeInstanceTypes
      tags: [compute]
      summary: List the instance sizes, smallest first.
      description: The same names Terraform uses in platform/terraform/flavors.yaml.
      responses:
        '200':
          description: Instance types.
          content:
            application/json:
              schema:
                type: object
                required: [instance_types]
                properties:
                  instance_types:
                    type: array
                    items:
                      type: object
                      required: [instance_type, vcpus, memory_mib, memory_min_mib]
                      properties:
                        instance_type: {type: string}
                        vcpus: {type: integer}
                        memory_mib:
                          type: integer
                          description: The ceiling, and what quota and capacity count.
                        memory_min_mib:
                          type: integer
                          description: The floor the host may balloon down to.
        '401': {$ref: '#/components/responses/Error'}
  /v1/capacity:
    get:
      operationId: DescribeCapacity
      tags: [compute]
      summary: Report what the host has, what the cloud has handed out, and the limits in force.
      description: |
        Answers the question a caller has before launching something large: is
        there room, and am I allowed.

        **Allotted figures are sums of ceilings, not measurements.** Memory is
        ballooned, so a guest may be holding far less than its `memory_mib`.
        The node's own figures and the allotted total will not agree, and both
        are reported. cloud-admins also get a per-account breakdown.
      responses:
        '200':
          description: Capacity.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capacity'}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/limits:
    get:
      operationId: DescribeLimits
      tags: [compute]
      summary: Read the limits in force, the deployment's defaults, and what an administrator changed.
      responses:
        '200':
          description: Limits.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LimitsEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    put:
      operationId: UpdateLimits
      tags: [compute]
      summary: Change the limits. cloud-admins only.
      description: |
        **Replaces every override at once.** A field left out of the request
        goes back to the deployment's default from
        `platform/terraform/cloud.yaml`, so an administrator cannot be left with
        a limit they have forgotten setting. Send `{}` to return everything to
        the defaults.

        **0 means unlimited** for the account quotas and `memory_budget_mib`,
        and turns `vm_disk_max_used_percent` off. `node_memory_reserve_mib` and
        `vm_disk_max_used_percent` are the checks that look at real capacity, so
        setting them to 0 removes the protection against filling the host; the
        request that the memory being asked for is actually free is never
        skipped.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LimitOverrides'}
      responses:
        '200':
          description: Stored.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LimitsEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
  /v1/instances/adopt:
    post:
      operationId: AdoptInstance
      tags: [compute]
      summary: Register a VM that already exists into the cloud. cloud-admins only.
      description: |
        cloud-admins only.

        Takes over a VM the administrator has already moved into the cloud
        pool. It is registered under `account_id` and from then on behaves like
        a launched instance: power, console, tags and security groups all work.
        It cannot be launched or re-created, so it keeps no source image or
        seed ISO.

        The VM must be running or stopped, have a net0 MAC address, and its
        root disk must be measurable; if storage does not report the size, pass
        `root_disk_gib`. An address is optional, but security groups filter by
        address, so an adopted VM with none cannot be filtered.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdoptInstanceRequest'}
      responses:
        '201':
          description: The adopted instance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
  /v1/instances/{instance_id}/start:
    post:
      operationId: StartInstance
      tags: [compute]
      summary: Start a stopped instance.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /v1/instances/{instance_id}/stop:
    post:
      operationId: StopInstance
      tags: [compute]
      summary: Stop a running instance.
      description: The guest is asked to shut down first, and forced after two minutes.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /v1/instances/{instance_id}/reboot:
    post:
      operationId: RebootInstance
      tags: [compute]
      summary: Reboot a running instance.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /v1/volumes/{volume_id}/detach:
    post:
      operationId: DetachVolume
      tags: [volumes]
      summary: Detach a volume from its instance.
      description: |
        Owner or cloud-admins only (403 UnauthorizedOperation otherwise).
        The volume must be `in-use` (409 IncorrectState otherwise). Detaching
        expects the guest to have already unmounted the device; the API does
        not ask the guest to unmount it first.

        Hot-unplugs from a running instance; the guest no longer sees the
        device once detached.
      parameters:
        - $ref: '#/components/parameters/VolumeID'
      responses:
        '202':
          description: Accepted; the volume is detaching.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VolumeEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
components:
...
  securitySchemes:
    accessKey:
      type: http
      scheme: bearer
      bearerFormat: sca_<access_key_id>.<secret>
    session:
      type: apiKey
      in: cookie
      name: shakecloud_session
...
  schemas:
...
    CallerIdentity:
      type: object
      required: [account_id, username, is_admin, credential_type]
      properties:
        account_id: {type: string, pattern: '^[0-9]{12}$'}
        username: {type: string}
        is_admin: {type: boolean}
        credential_type: {type: string, enum: [session, access_key]}
        access_key_id: {type: string}
...
    ImageEnvelope:
      type: object
      required: [image]
      properties:
        image: {$ref: '#/components/schemas/Image'}
...
    AccessKey:
      type: object
      required: [access_key_id, status, description, create_date]
      properties:
        access_key_id: {type: string, pattern: '^[a-z2-7]{20}$'}
        status:
          type: string
          enum: [Active, Inactive, Expired]
        description: {type: string}
        create_date: {type: string, format: date-time}
        expire_date: {type: string, format: date-time}
        last_used_date: {type: string, format: date-time}
...
    KeyPairEnvelope:
      type: object
      required: [key_pair]
      properties:
        key_pair: {$ref: '#/components/schemas/KeyPair'}
...